restructure(PageTreeModule): require name of PageTreeNode entity to be 'PageTreeNode' and remove createInternalLinkBlock factory
