+++ packages/api/cms-api/src/page-tree/page-tree.constants.ts
@@ -1,5 +1,7 @@
 export const PAGE_TREE_REPOSITORY = "PageTreeRepository";
 
+export const PAGE_TREE_ENTITY = "PageTreeNode";
+
 export const PAGE_TREE_CONFIG = "PageTreeConfig";
 
 export const defaultReservedPaths = ["/admin", "/preview"];
